requirements.txt
django==4.2.16
pillow
django-summernote
django-simple-captcha